backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pydantic==2.10.3

# LangChain ecosystem
langchain==0.3.13
langchain-openai==0.2.14
langchain-community==0.3.13
langgraph==0.2.60

# Tools
tavily-python==0.5.0
sendgrid==6.11.0

# Utilities
httpx==0.28.1
